requirement.txt
click==8.0.3
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
Flask==2.0.2
imageio==2.10.1
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.5
opencv-python==4.5.4.58
Pillow==8.4.0
pkg_resources==0.0.0
pyparsing==3.0.1
python-dateutil==2.8.2
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.4
six==1.16.0
tifffile==2020.9.3
typing-extensions==3.10.0.2
uWSGI==2.0.20
Werkzeug==2.0.2
zipp==3.6.0
